исправление по итогу код-ревью
