+++ 18-API/src/app.ts
@@ -16,7 +16,7 @@ import type { UserController } from './modules/users';
 import type { UserService } from './modules/users';
 import type { WarehouseBalancesController } from './modules/warehouse-balances/warehouse-balance.controller';
 import { TYPES } from './types';
-import { TelegramBot } from './modules/telegram/TelegramBot';
+import { TelegramBot } from './modules/telegram/telegram.bot';
 
 
 @injectable()
@@ -70,7 +70,7 @@ export class App {
 		this.useExeptionFilters();
 		await this.prismaService.connect();
 		this.server = this.app.listen(this.port);
-		this.telegramBot.bot.launch();
+		this.telegramBot.start();
 		this.logger.log(`Сервер запущен на http://localhost:${this.port}`);
 	}
