+++ 18-API/src/modules/telegram/enums/enums.ts
@@ -1,0 +1,5 @@
+export enum ETelegramSceneNames {
+    profile = 'profile',
+    items = 'items',
+    promo = 'promo',
+};
