+++ 18-API/src/modules/telegram/interfaces/ibot.interfaces.ts
@@ -1,3 +1,4 @@
 export interface ITelegramBot {
     init: () => void;
+    start: () => void;
 };
